Task 004 stage 5c: every tracking mode logs as itself, and the checkmark stops completing empty sets
The set row now draws what its exercise tracks (FR-2.3): weight x reps with
RIR; reps with RIR; a time alone; weight, distance and time for a carry. The
15 seeded holds and carries stop being logged as weight x reps, and the 31
reps-only exercises lose a weight field. A hold and a carry have no RIR - no
reps to hold in reserve - so it stays NULL (INV-03). A time is typed on the
app's keypad, filling from the right (130 is 1:30).

The checkmark no longer completes a row missing what its mode needs - 03
section 4's "enforced in the service layer", enforced at last by a pure
predicate in src/db. It opens the keypad on the missing field instead.

set_logs.distance_m becomes numeric(9,3) / real: as an integer, 100 ft read
back as 98 ft. The device migration is the rebuild drizzle-kit generated, kept
because it was read and proven first (06 section 4) - nothing references
set_logs on the device, and against a copy of the test phone's database every
set survived and both CHECKs still bit after the rename.

The create form offers all four modes; the targets sheet drops the rep range
and RIR for holds and carries and clears them on save; a routine start
pre-fills last time's time and distance.

Found on the way: "Duplicate routine" could never have worked - stage 5a
spread a whole routine exercise in as its targets, after the row's id, so
every copy reused the original's ids. Fixed with targetsOf() and a test. And
the INV-23 fence reads any `duration: <literal>` as an animation timing;
worked within it rather than changing its enforcement without an ADR.

Mobile: tsc, eslint, 49 lint fixtures, 630 catalog messages, no schema drift,
942 Jest tests. API: alembic up/down/up through 0006, check_schema, 347
pytest tests. Device pass not yet run.

+++ apps/api/app/models/strength.py
@@ -186,7 +186,8 @@ class SetLog(OwnedByUser, SyncColumns, Base):
     weight_kg: Mapped[Decimal | None] = mapped_column(Numeric(9, 4))
     reps: Mapped[int | None] = mapped_column(SmallInteger)
     rir: Mapped[int | None] = mapped_column(SmallInteger)
-    distance_m: Mapped[int | None] = mapped_column(Integer)
+    # Decimals, so a distance typed in feet reads back as typed (task 004 stage 5c, 03 §4).
+    distance_m: Mapped[Decimal | None] = mapped_column(Numeric(9, 3))
     duration_s: Mapped[int | None] = mapped_column(Integer)
     is_completed: Mapped[bool] = mapped_column(server_default="false")
     completed_at: Mapped[datetime | None]
